# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是一个简单的校园点评网站的后端 API 文档。",
        "title": "杭电点评 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/reviews": {
            "get": {
                "description": "获取已审核通过的点评列表，支持分页、搜索和排序。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点评"
                ],
                "summary": "公开点评列表",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "搜索关键词",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "-created_at",
                        "description": "排序字段 (e.g., -created_at, rating)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ReviewListResult"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "创建一个新的评价。如果店铺不存在，则根据提供的店铺信息自动创建。此接口为幂等操作。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "点评"
                ],
                "summary": "创建新评价（可自动创建店铺）",
                "parameters": [
                    {
                        "type": "string",
                        "description": "幂等键 (UUID)，用于防止重复提交",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "评价和新店铺信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateReviewForNewStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/dto.AutoCreateReviewResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "409": {
                        "description": "用户已评价过该店铺或请求正在处理中",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "429": {
                        "description": "请求正在处理中",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
...
        "/stores": {
            "get": {
                "description": "根据多种条件搜索已审核通过的店铺，支持分页、排序和过滤。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "店铺"
                ],
                "summary": "搜索店铺",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词 (名称或地址)",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "-created_at",
                        "description": "排序字段 (e.g., -created_at, average_rating)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "店铺分类",
                        "name": "category",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "店铺状态 (仅管理员可用)",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.StoreListResult"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "用户创建一个新的店铺，需要等待管理员审核。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "店铺"
                ],
                "summary": "创建新店铺",
                "parameters": [
                    {
                        "description": "店铺信息",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateStoreRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/dto.StoreResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "409": {
                        "description": "店铺已存在",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
        "/stores/{id}": {
            "get": {
                "description": "根据 ID 获取单个店铺的详细信息。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "店铺"
                ],
                "summary": "获取店铺详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店铺 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "店铺详情",
                        "schema": {
                            "$ref": "#/definitions/dto.StoreResponse"
                        }
                    },
                    "400": {
                        "description": "无效的店铺 ID",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "404": {
                        "description": "店铺不存在",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
        "/stores/{id}/reviews": {
            "get": {
                "description": "获取指定店铺下所有已审核通过的评价。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "店铺"
                ],
                "summary": "获取店铺的评价列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店铺 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "每页数量",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "default": "-created_at",
                        "description": "排序字段 (e.g., -created_at, rating)",
                        "name": "sort",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/services.ReviewListResult"
                        }
                    },
                    "400": {
                        "description": "无效的店铺 ID",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "为指定 ID 的店铺提交一条新评价。此接口为幂等操作。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "店铺"
                ],
                "summary": "为店铺提交评价",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店铺 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "幂等键 (UUID)，用于防止重复提交",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "评价内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/dto.ReviewResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "404": {
                        "description": "店铺不存在",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "409": {
                        "description": "用户已评价过该店铺或请求正在处理中",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
        "/stores/{id}/reviews/{reviewId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "删除用户对指定店铺的评价。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "店铺"
                ],
                "summary": "删除店铺评价",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店铺 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "评价 ID",
                        "name": "reviewId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "删除成功"
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "403": {
                        "description": "无权操作",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "404": {
                        "description": "评价不存在",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新用户对指定店铺的评价。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "店铺"
                ],
                "summary": "更新店铺评价",
                "parameters": [
                    {
                        "type": "string",
                        "description": "店铺 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "评价 ID",
                        "name": "reviewId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "要更新的评价内容",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateReviewRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "$ref": "#/definitions/dto.ReviewResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "401": {
                        "description": "未认证",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "403": {
                        "description": "无权操作",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    },
                    "404": {
                        "description": "评价不存在",
                        "schema": {
                            "$ref": "#/definitions/problem.Details"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "dto.AuthorResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.AutoCreateReviewResponse": {
            "type": "object",
            "properties": {
                "is_new_store": {
                    "type": "boolean"
                },
                "review": {
                    "$ref": "#/definitions/dto.ReviewResponse"
                },
                "store": {
                    "$ref": "#/definitions/dto.StoreResponse"
                }
            }
        },
        "dto.CreateReviewForNewStoreRequest": {
            "type": "object",
            "required": [
                "content",
                "rating",
                "store",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "rating": {
                    "type": "number",
                    "maximum": 5,
                    "minimum": 0
                },
                "store": {
                    "$ref": "#/definitions/dto.StoreInfoForReview"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.CreateReviewRequest": {
            "type": "object",
            "required": [
                "content",
                "rating",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "rating": {
                    "type": "number",
                    "maximum": 5,
                    "minimum": 0
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "dto.CreateStoreRequest": {
            "type": "object",
            "required": [
                "address",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "dto.ReviewResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "$ref": "#/definitions/dto.AuthorResponse"
                },
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ImageResponse"
                    }
                },
                "rating": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto.StoreInfoForReview": {
            "type": "object",
            "required": [
                "address",
                "name"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.StoreResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "average_rating": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "total_reviews": {
                    "type": "integer"
                }
            }
        },
        "dto.UpdateReviewRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "rating": {
                    "type": "number",
                    "maximum": 5,
                    "minimum": 0
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "problem.Details": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "instance": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "services.Pagination": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "services.ReviewListResult": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ReviewResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/services.Pagination"
                }
            }
        },
        "services.StoreListResult": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.StoreResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/services.Pagination"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}